backend/requirements.txt
# requirements.txt
fastapi
uvicorn
motor
python-dotenv
pymongo[srv]
pandas
scikit-learn
certifi
dnspython
requests